integrando end-point de finalização de servicos

+++ app/src/main/java/com/example/projetotcc/servico/data/repository/ServicoService.kt
@@ -26,7 +26,7 @@ interface ServicoService {
 
     @PUT("api/v1/servico/finalizar/{idTarefa}")
     suspend fun finalizarTarefaPorId(
-        @Path("idTarefa") idTarefa: Long,
+        @Path("idTarefa") idTarefa: Int,
     ): Response<ServicoResponse>
 
     @PATCH("api/v1/servico/custo/{idTarefa}")
